% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_transformations.R
\name{add_common_effect_sizes}
\alias{add_common_effect_sizes}
\title{Add common effect size columns to FReD dataset}
\usage{
add_common_effect_sizes(
  fred_data,
  es_value_columns = c("es_orig_value", "es_rep_value"),
  es_type_columns = c("es_orig_estype", "es_rep_estype"),
  es_common_names = c("es_original", "es_replication")
)
}
\arguments{
\item{fred_data}{FReD dataset}

\item{es_value_columns}{Character vector of column names for effect sizes}

\item{es_type_columns}{Character vector of column names for effect size types}
}
\value{
FReD dataset with additional columns for common effect sizes
}
\description{
Converts original and replication effect sizes to common metric (r) and adds them to the dataset.
}
