% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_transformations.R
\name{add_replication_power}
\alias{add_replication_power}
\title{Add power}
\usage{
add_replication_power(
  fred_data,
  es_original = "es_original",
  N_replication = "n_replication",
  power_column = "power_r"
)
}
\arguments{
\item{fred_data}{FReD dataset}

\item{es_original}{Character. Name of original effect size column.}

\item{power_column}{Character. Name of target column for power.}

\item{sample_replication}{Character. Name of replication sample size column.}
}
\value{
Augmented FReD dataset with power column.
}
\description{
Estimates the power of the replication study, given the original effect size, the sample size of the replication study, and the usual focus on a two-tailed test.
}
