% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_transformations.R
\name{add_uncertainty}
\alias{add_uncertainty}
\title{Add sampling variances, confidence intervals and p-values}
\usage{
add_uncertainty(
  fred_data,
  es_value_columns = c("es_original", "es_replication"),
  N_columns = c("n_original", "n_replication"),
  vi_columns = c("vi_original", "vi_replication"),
  ci_lower_columns = c("ci.lower_original", "ci.lower_replication"),
  ci_upper_columns = c("ci.upper_original", "ci.upper_replication"),
  p_values = c("p_value_original", "p_value_replication")
)
}
\arguments{
\item{fred_data}{FReD dataset}

\item{es_value_columns}{Character vector of column names with correlation values}

\item{N_columns}{Character vector of column names with sample sizes}

\item{se_columns}{Character vector of target columns for standard errors}
}
\value{
FReD dataset with additional columns for standard errors
}
\description{
Adds sampling variances, confidence intervals (asymmetric, using z-transformation) and p-values for common-metric effect sizes (r) to the FReD dataset.
}
