% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_transformations.R
\name{align_effect_direction}
\alias{align_effect_direction}
\title{Align effect direction}
\usage{
align_effect_direction(
  fred_data,
  es_original = "es_original",
  es_replication = "es_replication"
)
}
\arguments{
\item{fred_data}{FReD dataset}

\item{es_original}{Character. Name of original effect size column.}

\item{es_replication}{Character. Name of replication effect size column.}
}
\value{
Augmented FReD dataset with aligned effect directions.
}
\description{
Ensure that all original effects are coded as positive, and that replication effects are coded in the same direction (so that they \emph{would} be positive if successful.
}
