% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/assets.R
\docType{data}
\name{assets}
\alias{assets}
\title{Assets}
\format{Data frame with columns
\describe{
\item{period}{Date Weekly}
\item{assets}{numeric Assets on Fed balance sheet, millions USD}
}}
\source{
\url{http://www.federalreserve.gov/datadownload/Choose.aspx?rel=H41}
}
\usage{
assets
}
\description{
The US money supply M1 and M2 measures.
}
\examples{
data(assets)
}
\keyword{datasets}

