% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{bind_rows_with_characters}
\alias{bind_rows_with_characters}
\title{Bind Rows with Character Columns}
\usage{
bind_rows_with_characters(..., .id = NULL)
}
\arguments{
\item{...}{Data frames to combine. Each argument should be a data frame.}

\item{.id}{An optional string that will be used to create a column in the output.
If supplied, this will create a new column with the name given by \code{.id}, and each
row will have a value corresponding to the argument name from which it came.}
}
\value{
A data frame created by binding the rows of the input data frames.
}
\description{
A wrapper for \code{dplyr::bind_rows} that ensures any columns that are character
in one of the data frames are converted to character in all data frames before binding.
This reduces the likelihood of errors, but does not give up on type-checking entirely.
}
\keyword{internal}
