% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_prediction_interval}
\alias{calculate_prediction_interval}
\title{Calculate Prediction Interval for a Correlation Coefficient}
\usage{
calculate_prediction_interval(r, n_o, n_r = n_o, confidence_level = 0.95)
}
\arguments{
\item{r}{Numeric. The correlation coefficient from the original study.}

\item{n_o}{Integer. The sample size of the original study.}

\item{n_r}{Integer. The sample size of the replication study. Defaults to the same as \code{n_o}.}

\item{confidence_level}{Numeric. The confidence level for the prediction interval. Defaults to 0.95.}
}
\value{
A named numeric vector with two elements:
\item{lower}{The lower bound of the prediction interval.}
\item{upper}{The upper bound of the prediction interval.}
}
\description{
This function calculates the prediction interval for a correlation coefficient
based on the Fisher's Z-transformation. It uses the approach proposed in the
supplementary materials of Patil et al. (2016).
}
\examples{
# Example usage:
r <- 0.5   # Correlation coefficient
n_o <- 30  # Sample size of the original study
n_r <- 30  # Sample size of the replication study

calculate_prediction_interval(r, n_o, n_r)
}
\references{
Patil, P., Peng, R. D., & Leek, J. T. (2016). A statistical framework for enhancing
reproducibility in data analysis. \emph{Perspectives on Psychological Science, 11}(5), 633-640.
\url{https://doi.org/10.1177/1745691616646366}
}
