% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/capacity.R
\docType{data}
\name{capacity}
\alias{capacity}
\title{Capacity}
\format{Data frame with columns
\describe{
\item{period}{Date monthly}
\item{utilization}{numeric Industrial capacity utilization (percent)}
\item{capacity}{numeric Industrial capacity index}
\item{durable}{numeric Durable goods capacity index}
\item{nondurable}{numeric Nondurable goods capacity index}
}}
\source{
\url{http://www.federalreserve.gov/datadownload/Choose.aspx?rel=G17}
}
\usage{
capacity
}
\description{
US industrial capacity.
}
\examples{
data(capacity)
}
\keyword{datasets}

