% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{clean_variables}
\alias{clean_variables}
\title{Clean variables
Perform some specific operations (e.g., recoding some NA as "") required to get the Shiny apps to work.
This may be a temporary solution, as much of it should likely be handled through validation in the data sheet, and at import time.}
\usage{
clean_variables(fred_data)
}
\arguments{
\item{fred_data}{FReD dataset}
}
\value{
FReD dataset with cleaned variables
}
\description{
Clean variables
Perform some specific operations (e.g., recoding some NA as "") required to get the Shiny apps to work.
This may be a temporary solution, as much of it should likely be handled through validation in the data sheet, and at import time.
}
