% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_transformations.R
\name{code_replication_outcomes}
\alias{code_replication_outcomes}
\title{Code replication outcomes}
\usage{
code_replication_outcomes(
  fred_data,
  es_original = "es_original",
  p_original = "p_value_original",
  p_replication = "p_value_replication",
  ci_lower_replication = "ci.lower_replication",
  ci_upper_replication = "ci.upper_replication",
  es_replication = "es_replication"
)
}
\arguments{
\item{fred_data}{FReD dataset}

\item{es_original}{Character. Name of original effect size column.}

\item{p_original}{Character. Significance of original effect size.}

\item{p_replication}{Character. Significance of replication effect size.}

\item{ci_lower_replication}{Character. Lower bound of replication confidence interval.}

\item{ci_upper_replication}{Character. Upper bound of replication confidence interval.}

\item{es_replication}{Character. Name of replication effect size column.}
}
\value{
Augmented FReD dataset with replication outcome columns, including \code{signal}
}
\description{
Different frameworks have been proposed to code replication outcomes. Here we code:
\itemize{
\item signal vs no-signal (i.e., significant vs all others)
\item consistent vs inconsistent (i.e., replication confidence interval overlaps original point estimate)
Based on https://etiennelebel.com/documents/lebeletal\%282018,ampss\%29a-unified-framework-to-quantify-the-credibility-of-scientific-findings.pdf
\item and success vs failure (significant \emph{in right direction} vs all others)
}
}
