% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_transformations.R
\name{convert_effect_sizes}
\alias{convert_effect_sizes}
\title{Convert effect sizes to common metric (r)}
\usage{
convert_effect_sizes(es_values, es_types)
}
\arguments{
\item{es_values}{Numeric vector of effect sizes}

\item{es_types}{Character vector of effect size types (types/wordings that are not supported are flagged in warning)}
}
\value{
Numeric vector of effect sizes in common metric (r)
}
\description{
Takes vectors of effect sizes and their types and converts them to a common metric (r).
}
