% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{create_citation}
\alias{create_citation}
\title{Create FReD dataset citation}
\usage{
create_citation(data_file = get_param("FRED_DATA_FILE"), cache = TRUE)
}
\arguments{
\item{data_file}{Path to the FReD dataset, defaults to the current FReD dataset on OSF}

\item{cache}{Should the citation be returned from cache, if already requested during this session? Defaults to TRUE.}
}
\value{
A markdown-formatted citation for the FReD dataset, including the current dataset version.
}
\description{
Pulls current contributor list and dynamicalky creates a \emph{markdown-formatted} citation for the FReD dataset.
}
