% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/credit.R
\docType{data}
\name{credit}
\alias{credit}
\title{Credit}
\format{Data frame with columns
\describe{
\item{period}{Date Monthly}
\item{nonrevolving}{numeric Nonrevolving credit}
\item{revolving}{numeric Revolving credit}
\item{student}{numeric Student loans}
}}
\source{
\url{http://www.federalreserve.gov/datadownload/Choose.aspx?rel=G19}
}
\usage{
credit
}
\description{
Outstanding consumer credit balances in millions of USD.
}
\examples{
data(credit)
}
\keyword{datasets}

