% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_parameters.R
\name{.onLoad}
\alias{.onLoad}
\title{Setting parameters for the FReD package}
\usage{
.onLoad(libname, pkgname)
}
\description{
Generally, the FReD package is designed to work with the latest FReD dataset.
However, you might want to use an older version, or even your own. For such advanced
use, you can set parameters, either using \code{Sys.setenv(FRED_DATA_URL = "http://your_url")}
or by setting \code{options(FRED_DATA_URL = "http://your_url")}. The following parameters
can be set (before loading the package):
}
\details{
\itemize{
\item \code{FRED_DATA_URL}: The URL of the FReD dataset, needs to return the .xlsx file.
\item \code{FRED_DATA_FILE}: The path to the .xlsx file, if you have downloaded it already (or want it to be saved to a particular location)
}
}
