% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot data accident}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{an integer that represent the number of a state.}

\item{year}{an integer that represent a year.}
}
\value{
This function returns a graph where can be seen the location (latitude and
longitude) of the accident that happend in the state in the year entered.
}
\description{
This functions allow to plot the longitude and latitude of the accidents that
happend in a specific state in a specific year
}
\note{
This function will generate an error if the year or state
does not exists in the database.
}
\examples{
fars_map_state(27,2013)
}
