% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a compressed file.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A caracter that contains the name of the file with its respective
extension.}
}
\value{
This function returns the read data in a "dataframe", "tbl_df", "tbl" format.
}
\description{
This functions allow to read a compressed (.bz2) or a csv file.
}
\note{
This function will generate an error if the filemane is wrong or have a
wrong extension.
}
\examples{
library(dplyr)
library(readr)
fars_read(make_filename(2013))
fars_read(make_filename(2014))
}
