% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Select specific variables of the data.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{an integer that represent a year or years.}
}
\value{
This function returns a list with the variable "MONTH" and "YEAR" for the
year or years entered.
}
\description{
This functions allow to select the variable "MONTH" and "YEAR" from
the data for the accidents that happend in the year or years entered.
}
\note{
This function will generate an warning message if the year or years entered
does not exists in the database.
}
\examples{
fars_read_years(c(2013,2014))
fars_read_years(2013)
}
