% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Sumarize accidents by month of a specific year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{an integer that represent a year or years.}
}
\value{
This function returns a summary for every month of every year
in a "dataframe", "tbl_df", "tbl" format.
}
\description{
This functions allow to generate a summary of the accidents that happend in a
specific year or years.
}
\note{
This function will generate an warning message if the year or years entered
does not exists in the database.
}
\examples{
fars_summarize_years(c(2013,2014))
fars_summarize_years(c(2013,2015))
}
