% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categories.R
\name{fred_category_series}
\alias{fred_category_series}
\title{List series in a category}
\usage{
fred_category_series(category_id, limit = 1000L)
}
\arguments{
\item{category_id}{Integer. The category ID.}

\item{limit}{Integer. Maximum number of results to return. Default 1000.}
}
\value{
A data frame of series metadata.
}
\description{
Returns all series belonging to a given category. Automatically paginates
through all results.
}
\examples{
\donttest{
op <- options(fred.cache_dir = tempdir())
fred_category_series(32992)
options(op)
}
}
\seealso{
Other categories: 
\code{\link{fred_category}()},
\code{\link{fred_category_children}()}
}
\concept{categories}
