% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{fred_get_key}
\alias{fred_get_key}
\title{Get the current FRED API key}
\usage{
fred_get_key()
}
\value{
Character. The API key.
}
\description{
Returns the API key set via \code{\link[=fred_set_key]{fred_set_key()}} or the \code{FRED_API_KEY}
environment variable. Raises an error if no key is found.
}
\examples{
\donttest{
op <- options(fred.cache_dir = tempdir())
fred_get_key()
options(op)
}
}
\seealso{
Other configuration: 
\code{\link{clear_cache}()},
\code{\link{fred_request}()},
\code{\link{fred_set_key}()}
}
\concept{configuration}
