% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{fred_related_tags}
\alias{fred_related_tags}
\title{Find tags related to a given tag}
\usage{
fred_related_tags(tag_names)
}
\arguments{
\item{tag_names}{Character. One or more tag names, separated by semicolons
(e.g. \code{"gdp"}, \code{"usa;quarterly"}).}
}
\value{
A data frame of related tags.
}
\description{
Returns tags that are frequently used together with the specified tag.
}
\examples{
\donttest{
op <- options(fred.cache_dir = tempdir())
fred_related_tags("gdp")
options(op)
}
}
\seealso{
Other tags: 
\code{\link{fred_tags}()}
}
\concept{tags}
