% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{fred_release_dates}
\alias{fred_release_dates}
\title{Get release dates}
\usage{
fred_release_dates(release_id)
}
\arguments{
\item{release_id}{Integer. The release ID.}
}
\value{
A data frame with columns \code{release_id} and \code{date}.
}
\description{
Returns the dates on which data for a release were published. Useful for
understanding the data calendar and when revisions occurred.
}
\examples{
\dontrun{
fred_release_dates(53)
}
}
\seealso{
Other releases: 
\code{\link{fred_release_series}()},
\code{\link{fred_releases}()}
}
\concept{releases}
