% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{fred_release_series}
\alias{fred_release_series}
\title{List series in a release}
\usage{
fred_release_series(release_id)
}
\arguments{
\item{release_id}{Integer. The release ID.}
}
\value{
A data frame of series metadata.
}
\description{
Returns all series belonging to a given release.
}
\examples{
\dontrun{
# GDP release
fred_release_series(53)
}
}
