% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sources.R
\name{fred_source_releases}
\alias{fred_source_releases}
\title{List releases from a source}
\usage{
fred_source_releases(source_id)
}
\arguments{
\item{source_id}{Integer. The source ID.}
}
\value{
A data frame of releases.
}
\description{
Returns all releases published by a given data source.
}
\examples{
\donttest{
op <- options(fred.cache_dir = tempdir())
# Bureau of Labor Statistics
fred_source_releases(22)
options(op)
}
}
\seealso{
Other sources: 
\code{\link{fred_sources}()}
}
\concept{sources}
