% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.R
\name{fred_updates}
\alias{fred_updates}
\title{List recently updated FRED series}
\usage{
fred_updates(limit = 100L)
}
\arguments{
\item{limit}{Integer. Maximum number of results. Default 100, maximum 100.}
}
\value{
A data frame of recently updated series.
}
\description{
Returns series that have been recently updated or revised.
}
\examples{
\dontrun{
fred_updates()
}
}
