% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{get_dataset_changelog}
\alias{get_dataset_changelog}
\title{Get the dataset changelog from OSF}
\usage{
get_dataset_changelog(
  changelog_file = "https://osf.io/fj3xc/download",
  cache = TRUE
)
}
\arguments{
\item{changelog_file}{The URL of the changelog file on OSF}

\item{cache}{Should the changelog be returned from cache, if already requested during this session? Defaults to TRUE.}
}
\description{
Downloads and reads the changelog file from OSF, and returns it as a character string.
}
