% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fred.R
\name{load_fred_data}
\alias{load_fred_data}
\title{Load the FReD dataset}
\usage{
load_fred_data(data = get_param("FRED_DATA_FILE"))
}
\arguments{
\item{data}{Path to the FReD dataset (defaults to current FReD data on OSF)}
}
\value{
A data frame with the processed FReD dataset
}
\description{
This function loads the FReD dataset into R, and conducts variable transformations to  prepare for analyses.
}
