% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{load_variable_descriptions}
\alias{load_variable_descriptions}
\title{Load variable descriptions}
\usage{
load_variable_descriptions(
  sheet_name = "Key Variables",
  data = get_param("FRED_DATA_FILE")
)
}
\arguments{
\item{sheet_name}{Path to the variable descriptions}

\item{data}{Path to the FReD dataset (defaults to current FReD data on OSF)}
}
\value{
A data frame with variable names (\code{Variable}) and descriptions (\code{Description})
}
\description{
This reads names and variable descriptions of key variables from FReD. This can used for subsetting and describing the dataset.
}
