% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{an integer that represent a year.}
}
\value{
This function returns a caracter with a filename.
}
\description{
This functions allow to create a character with a filename, that will have a
"csv.bz2" extension. The filename will always be "accident_*.csv.bz2", with
the only diference that "*" will be replace by the entered year.
}
\note{
This function will generate an warning message if the parameter
"year" is a character instead a integer.
}
\examples{
make_filename(2013)
make_filename(2014)
}
