% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/money.R
\docType{data}
\name{money}
\alias{money}
\title{Money}
\format{Data frame with columns
\describe{
\item{period}{Date Weekly}
\item{m1}{numeric Total M1 supply}
\item{currency}{numeric Currency in circulation}
\item{demand}{numeric Demand deposits}
\item{other}{numeric Other short-term deposits}
\item{travelers}{numeric Traveler's cheques}
\item{nonM1M2}{numeric M2 measures not included in M1}
}}
\source{
\url{http://www.federalreserve.gov/datadownload/Choose.aspx?rel=H6}
}
\usage{
money
}
\description{
The US money supply M1 and M2 measures in millions of USD.
}
\examples{
data(money)
}
\keyword{datasets}

