% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_transformations.R
\name{p_from_r}
\alias{p_from_r}
\title{Calculate p-values from correlation coefficients and sample size}
\usage{
p_from_r(r, N)
}
\arguments{
\item{r}{Correlation coefficient.}

\item{N}{Sample size.}
}
\value{
Numeric p-value for the two-tailed test of the correlation.
}
\description{
This function calculates the p-value associated with a given correlation
coefficient and sample size, assuming the null hypothesis that the true
population correlation is zero.
}
\examples{
p_from_r(r = c(0.5, 0.3), N = c(30, 25))
}
