% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{print.fred_equivalence_test_result}
\alias{print.fred_equivalence_test_result}
\title{Print Method for Equivalence Test Result}
\usage{
\method{print}{fred_equivalence_test_result}(x, ...)
}
\arguments{
\item{x}{An object of class \code{equivalence_test_result}.}

\item{...}{Additional arguments (not used).}
}
\description{
This function is a custom print method for objects of class \code{equivalence_test_result}.
It prints a formatted sentence describing the results of the equivalence test.
}
