% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifest.R
\name{print.fred_manifest}
\alias{print.fred_manifest}
\title{Print method for fred_manifest}
\usage{
\method{print}{fred_manifest}(x, ...)
}
\arguments{
\item{x}{A \code{fred_manifest} object.}

\item{...}{Ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print method for fred_manifest
}
