% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{read_fred}
\alias{read_fred}
\title{Load the FReD dataset}
\usage{
read_fred(data = get_param("FRED_DATA_FILE"))
}
\arguments{
\item{data}{Path to the FReD dataset (defaults to current FReD data on OSF)}
}
\value{
A data frame with the FReD dataset
}
\description{
This function loads the FReD dataset into R. It merges the data from the different sheets into one data frame.
}
