% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps.R
\name{run_app}
\alias{run_app}
\title{Run a shiny app within the package - potentially as an RStudio job}
\usage{
run_app(
  offer_install = interactive(),
  app = "fred_explorer",
  in_background = NULL,
  port = 3838
)
}
\arguments{
\item{offer_install}{Should user be prompted to install required packages if they are missing?}

\item{app}{The name of the app folder within the package (i.e. within the inst folder)}

\item{in_background}{Should the app be run in the background (i.e. not block the R console)? Default to TRUE if RStudio is used.}

\item{port}{The port to run the app on (can usually be left at the default value)}
}
\value{
A shiny app
}
\description{
This function is used to run the shiny apps within the package
}
