% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps.R
\name{run_explorer}
\alias{run_explorer}
\title{Run the Replication Explorer}
\usage{
run_explorer(offer_install = interactive())
}
\arguments{
\item{offer_install}{Should user be prompted to install required packages if they are missing?}
}
\value{
Replication Explorer shiny app
}
\description{
Running this function will launch the FReD Replication Explorer shiny app
}
\examples{
if (interactive()) {
  # To run the CReplication Explorer app:
  run_explorer()
}
}
