% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{test_equivalence_r}
\alias{test_equivalence_r}
\title{Equivalence Test for a Correlation}
\usage{
test_equivalence_r(r, n, sesoi, alpha = 0.05)
}
\arguments{
\item{r}{Numeric. The observed correlation.}

\item{n}{Integer. The sample size.}

\item{sesoi}{Numeric. The smallest effect size of interest (SESOI) for the correlation.
The SESOI is symmetric around zero, so the function tests equivalence within the range
\code{-sesoi} to \code{sesoi}.}

\item{alpha}{Numeric. The significance level for the test. Default is 0.05.}
}
\value{
Numeric. The p-value for the equivalence test. A p-value less than the significance
level \code{alpha} indicates that the observed correlation is equivalent to the SESOI.
When the return value is printed, a full sentence detailing the result is provided.
}
\description{
This function performs an equivalence test for a given correlation \( r \) against a
specified smallest effect size of interest (SESOI). The test assesses whether the
observed correlation \( r \) is statistically equivalent to a range defined by the SESOI.
}
\details{
The logic behind the test is based on the Two One-Sided Tests (TOST) procedure.
The function tests two null hypotheses: (1) that the true correlation is less than or
equal to the lower bound of the SESOI, and (2) that the true correlation is greater than
or equal to the upper bound of the SESOI. If both null hypotheses can be rejected
(i.e., if the p-value is less than the significance level \code{alpha}), it is concluded
that the correlation is equivalent to the SESOI.
}
\examples{
r <- 0.3
n <- 100
sesoi <- 0.1
alpha <- 0.05
test_equivalence_r(r, n, sesoi, alpha)

}
