% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_parameters.R
\name{use_FReD_offline}
\alias{use_FReD_offline}
\title{Set FReD to work offline (or back to online)}
\usage{
use_FReD_offline(state = TRUE)
}
\arguments{
\item{state}{Should FReD work offline (TRUE) or online (FALSE)}
}
\description{
By default, FReD loads the latest data and meta-data every time it is loaded.
If you work offline, or don't need the latest data, you may prefer to work from
cached data. For that, you can call 'use_FReD_offline()'. If you want this to
persist even when FReD is reloaded, you can use \code{options(FRED_OFFLINE = TRUE)}
(or FALSE) as needed, and include it in your \code{.Rprofile} file to persist across
sessions.
}
