% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_transformations.R
\docType{data}
\name{z_test_data}
\alias{z_test_data}
\title{Augment Data for Z-Curve Analysis}
\format{
An object of class \code{data.frame} with 3 rows and 3 columns.
}
\usage{
z_test_data
}
\arguments{
\item{fred_data}{A dataframe containing \code{es_original}
for the effect size of the original study, and \code{n_original} for the
sample size of the original study.}
}
\value{
\code{fred_data} but with two additional columns:
\code{se} for the standard error, and \code{z} for the z-score.
}
\description{
This function calculates and appends the standard error and z-score
for z-curve analysis based on original effect size and sample size.
}
\keyword{datasets}
