% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effect_size_transformations.R
\name{add_sampling_variances}
\alias{add_sampling_variances}
\title{Add confidence intervals}
\usage{
add_sampling_variances(
  fred_data,
  es_value_columns = c("es_o", "es_r"),
  N_columns = c("n_o", "n_r"),
  vi_columns = c("vi_o", "vi_r")
)
}
\arguments{
\item{fred_data}{FReD dataset}

\item{es_value_columns}{Character vector of column names with correlation values}

\item{N_columns}{Character vector of column names with sample sizes}

\item{vi_columns}{Character vector of target columns for sampling variances}
}
\value{
FReD dataset with additional columns for sampling variances (metafor's \code{vi})
}
\description{
Adds sampling variances for common-metric effect sizes (r) to the FReD dataset, using metafor::escalc
}
