% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear the fred cache}
\usage{
clear_cache()
}
\value{
Invisible \code{NULL}.
}
\description{
Deletes all locally cached FRED data files. The next call to any data
function will re-download from the FRED API.
}
\examples{
\dontrun{
clear_cache()
}
}
\seealso{
Other configuration: 
\code{\link{fred_get_key}()},
\code{\link{fred_request}()},
\code{\link{fred_set_key}()}
}
\concept{configuration}
