% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_management.R
\name{read_fred}
\alias{read_fred}
\title{Read the FReD dataset}
\usage{
read_fred(
  data = get_param("FRED_DATA_FILE"),
  retain_es_as_character = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{data}{Path to the FReD dataset (defaults to current FReD data on OSF), unless the package is in offline mode (\code{use_FReD_offline()})}

\item{retain_es_as_character}{Should effect sizes be retained as character? Defaults to TRUE, so that coded test statistics with df can be converted to common metric.}

\item{verbose}{Should detailed messages be printed that highlight data conversion issues? Defaults to TRUE. FALSE is quiet mode, and NULL prints a summary of problems.}
}
\value{
A data frame with the FReD dataset
}
\description{
This function loads the FReD dataset into R. It merges the data from the different sheets into one data frame.
}
