% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_apps.R
\name{run_annotator}
\alias{run_annotator}
\title{Run the Replication Annotator}
\usage{
run_annotator()
}
\value{
Opens the web annotator in the default browser (invisibly returns NULL).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function previously launched a local Shiny app but now opens the
hosted web version of the FReD Annotator.
}
\examples{
if (interactive()) {
  run_annotator()
}
}
